+++ activerecord/CHANGELOG.md
@@ -1,3 +1,9 @@
+* Passing an Active Record object to `find` is now deprecated.  Call `.id`
+  on the object first.
+
+* Passing an Active Record object to `exists?` is now deprecated.  Call `.id`
+  on the object first.
+
 *   Only use BINARY for mysql case sensitive uniqueness check when column has a case insensitive collation.
 
     *Ryuta Kamizono*